+++ variables.tf
@@ -47,6 +47,18 @@ variable "tamr_instance_subnet" {
   description = "subnetwork to attach instance too"
 }
 
+variable "tamr_instance_deletion_protection" {
+  type        = bool
+  description = "Enabled deletion protection for the tamr VM"
+  default     = true
+}
+
+variable "tamr_external_ip" {
+  type        = bool
+  description = "Create and attach an external ip to tamr VM"
+  default     = false
+}
+
 variable "tamr_instance_project" {
   type        = string
   description = "The project to launch the tamr VM instance in."